json: correct C99 semantics and UTF-16 decoding (resolve #1)
The function `encode_utf8` accepted an `unsigned` type as its argument,
which per any C specification is at least two bytes. On some
implementations, this may not be enough to store a 21-bit code point.
This funciton also failed to encode any code points >= 0x010000 because
of a typo. Finally, this function would attempt to output characters
between 0xd800 and 0xdfff, which are high / low trail surrogates, not
characters.

The `read_unicode` function failed to check whether the result of
decoding an escape sequence was to yield a surrogate pair, meaning it
was impossible for `encode_utf8` to get a codepoint larger than 0xffff
anyway.

To make this easier, strtoll is avoided and the code point is calculated
while scanning hex characters.