json: allow custom allocator; improve stack semantics
Allocating and freeing a bunch of tiny structures for a stack and
strings can be costly from a performance perspective (as can relying on
the sytem allocator). This change allows consumers to specify their own
allocation functions.

This change also removes unused / unnecessary fields from the "nesting"
struct, and renames it to `struct json_stack`. This struct is now
embedded in `json_stream`, and functions as an array. The stack is grown
four elements at a time (which is ~1 cache line on many modern
architectures) because magic numbers are hard. The stack is now treated
as an array with a size and a pointer to the stack top, instead of being
a linked list.

When exiting from a particular nesting, we do not free or realloc to a
smaller size. Most JSON streams I've encountered are not terribly deep,
and those that are tend to have a relatively constant depth. In other
words, deep JSON data structures likely have pessimistic allocation
patterns; it's best to just reuse the memory. At 16 bytes per nesting
level, you really have to try hard to make this actually be a waste of
memory.

I've also reordered the `json_stream` structure a bit to try to keep
related data close. On architectures with 64B cache lines, `json_stream`
now takes 4, but all the common mutable data is in 1 while the read-only
stuff is elsewhere.