json: fix invalid behavior per JSONTestSuite
After this commit, the test suite at

https://github.com/nst/JSONTestSuite

reports no issues that aren't understood to be implementation-defined
behavior, meaning pdjson is the only C JSON parser that passes all tests
it is supposed to pass, and fails all tests it is supposed to fail. (At
least of the ones tested by JSONTestSuite.)

There remains ambiguity in various unspecified behaviors largely around
handling of unescaped UTF literals that are invalid or denote
surrogates, and various issues with subnormal numbers,
overflow/underflow, and issues of that ilk. In general, I would like to
error on those issues, but they aren't really pressing to me.

This commit introduces a "strict" mode which causes an error when
non-whitespace data trails the outermost value. This is not the default
mode because it prevents a consumer from parsing a stream of JSON with
no whitespace between the containers.