Add json_get_context() API function
This function returns the current parsing context, that is, JSON_OBJECT if we
are inside an object, JSON_ARRAY if we are inside an array, and JSON_DONE if
we are not yet/anymore in either.

Additionally, for the first two cases, it also returns the number of parsing
events that have already been observed at this level with json_next/peek().
In particular, inside an object, an odd number would indicate that the just
observed JSON_STRING event is a property name.