Improve the portability to platforms missing libc FILE streams
On embedded platforms, or in context such as bootloaders or kernels,
the FILE stream API is often not available. To solve this we first
split the header in two part, "pdjson_base.h" which contains all the
core implementation and "pdjson.h" which contains the platform
specific parts. This allow an external project that want to use pdjson
to provide its own "pdjson.h" were it can provide missing types or
additional platform specific interfaces.

After this we split the stream code to its own file and implement it
through the json_user_io API to also remove the FILE pointer from the
json_source struct.