Add support for using staticaly allocated memory
On platforms with limited memory it can be desirable to not use malloc
even if it is available. Add a function to allow using a statically
allocated stack and string buffer. When in this mode the allocator is
simply disabled so pdjson will error out if there is not enough memory.